populating slackNotifier with standard implementations of SlackClient, MessageFactory and ConfigurationFactory
