+++ TfsNotificationRelay.Slack/ISlackMessageFactory.cs
@@ -1,10 +1,9 @@
 namespace DevCore.TfsNotificationRelay.Slack
 {
-    using Configuration;
     using Notifications;
 
     public interface ISlackMessageFactory
     {
-        Message GetMessage(INotification notification, BotElement botElement, string channel);
+        Message GetMessage(INotification notification, SlackConfiguration slackConfiguration, string channel);
     }
 }
